builder refactor
