+++ src/taipy/gui/builder/__init__.py
@@ -8,3 +8,8 @@
 # Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
 # an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
+
+from ..renderers import ClassApi as Page
+from ..renderers._class_api.element_api_generator import _ElementApiGenerator
+
+_ElementApiGenerator().add_default()
