+++ src/taipy/gui/utils/__init__.py
@@ -17,7 +17,6 @@ from ._attributes import (
     _setscopeattr,
     _setscopeattr_drill,
 )
-from ._element_api_context_manager import _ElementApiContextManager
 from ._locals_context import _LocalsContext
 from ._map_dict import _MapDict
 from ._runtime_manager import _RuntimeManager
